refactor: move shell escape utilities to shared package
Move shellEscape(), buildShellCommand(), and SAFE_SHELL_RE from
agent/src/terminal-server.ts to shared/src/shell.ts so they can be
reused by other packages (e.g., server/src/ssh-manager.ts).
